Created pythagAddition.py
This is an off shoot of the feeInchesCalc.py that calculates the C Leg of a right triangle using the Pythagorean theorem. I needed a calculator that would help me keep a running total of Reinforced Concrete Gradebeams for a Custom Home estimate I was completing a takeoff for and all of the Grade Beams were C legs of Right triangles where the Architects only gave dimensions for the A & B legs of the triangles.